on line 22 I changed the
'cmd.Parameters.Add("@date", SqlDbType.DateTime).Value = movie.Date.ToString("yyyy-MM-dd"); '

to

'cmd.Parameters.Add("@date", SqlDbType.DateTime).Value = movie.Date.ToString("yyyy-MM-dd");'

This worked because the first code snippet, a parameter named "@date" is added to the SqlCommand object's Parameters collection. The parameter is explicitly defined with a data type of SqlDbType.DateTime. The value assigned to this parameter is the formatted string representation of the "date" variable, using the "yyyy-MM-dd HH:mm:ss" format. This approach explicitly specifies the data type of the parameter and converts the date to a string representation before assigning it.

The second code snippet, a parameter named "@date" is added to the SqlCommand object's Parameters collection using the AddWithValue method. The parameter value is directly set to the "movie.Date" property value, which is assumed to be of type DateTime. The AddWithValue method infers the data type of the parameter based on the provided value. In this case, the DateTime value is used directly without any explicit formatting or conversion.