Added display to editmovie. Ran into issue where image was showing broken image icon. Fixed two things. First fixed isssue where image type was not showing in the html tag. EX: <img src="data:image/;base64," alt="Image">
Fixed this by adding to the type to the movie Images model. Then I edited the code in the DataAccessLayer to be able to add and retieve that data from the database. Once that was done I was able to set the type using

"
var imgSrc = String.Format("data:image/{0};base64,{1}", Model.images[0].MimeType, base64);
"

Finally I had an issue where the image wasn't showing up. There was an empty image in the database so images weren't being properly uploaded to the database. This issue was present In the image utility folder right before uploading it to the database

I start off by declaring a byte array named imageData. This is where I'm going to store the content of the uploaded file.

Then, I want to make sure I'm reading the file content from the very beginning, so I set the Position of the InputStream to 0. The InputStream is where the uploaded file's data is stored, and Position tells me where in that stream I'm currently at. By setting it to 0, I'm making sure I start reading the file from the beginning.

Next, I create a BinaryReader by passing the InputStream to its constructor. This BinaryReader is going to help me read bytes from the InputStream.

I then use the ReadBytes method of the BinaryReader to read a number of bytes equal to the content length of the file from the InputStream. I store these bytes in imageData.

By adding the line file.InputStream.Position = 0;, I made sure that I was reading the file data from the start of the file. If the Position of the InputStream was not at the start for any reason, I would not have read the full content of the file, leading to an incomplete imageData. But by ensuring I always start reading from the beginning, I can be sure that I'm getting all the file data into imageData.